TW-1685: change from bytes to stream when downloading files web
(cherry picked from commit 92bbb6efea2b18074b6c4a9b2813bacb43741c95)
